[Migration 5-1] Feature: chain-variants/celo epoch pages
Closes #3434
