+++ client/features/chain-variants/celo/pages/epoch-details/__screenshots__/Epoch.pw.tsx_default_base-view-mobile-1.png
