+++ client/features/chain-variants/celo/pages/epoch-details/__screenshots__/Epoch.pw.tsx_default_unfinalized-epoch-1.png
