+++ client/features/chain-variants/celo/pages/epoch-details/__screenshots__/Epoch.pw.tsx_mobile_base-view-mobile-1.png
