+++ client/features/chain-variants/celo/pages/epoch-details/election-rewards/__screenshots__/EpochElectionRewards.pw.tsx_default_base-view-1.png
