+++ client/features/chain-variants/celo/pages/epoch-details/election-rewards/__screenshots__/EpochElectionRewards.pw.tsx_mobile_base-view-mobile---default-1.png
