+++ client/features/chain-variants/celo/pages/epoch-index/EpochsTable.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import type { CeloEpochListItem } from 'types/api/epochs';
+import type { CeloEpochListItem } from 'client/features/chain-variants/celo/types/api';
 
 import config from 'configs/app';
 import { TableBody, TableColumnHeader, TableHeaderSticky, TableRoot, TableRow } from 'toolkit/chakra/table';
