+++ client/features/chain-variants/celo/pages/epoch-index/__screenshots__/Epochs.pw.tsx_default_base-view-mobile-1.png
