+++ pages/epochs/[number].tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import type { Props } from 'nextjs/getServerSideProps/handlers';
 import PageNextJs from 'nextjs/PageNextJs';
 
-const Epoch = dynamic(() => import('ui/pages/Epoch'), { ssr: false });
+const Epoch = dynamic(() => import('client/features/chain-variants/celo/pages/epoch-details/Epoch'), { ssr: false });
 
 const Page: NextPage<Props> = (props: Props) => {
   return (
