+++ pages/epochs/index.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 
 import PageNextJs from 'nextjs/PageNextJs';
 
-const Epochs = dynamic(() => import('ui/pages/Epochs'), { ssr: false });
+const Epochs = dynamic(() => import('client/features/chain-variants/celo/pages/epoch-index/Epochs'), { ssr: false });
 
 const Page: NextPage = () => {
   return (
